requirements.txt
streamlit
biopython
py3Dmol
pandas
requests
matplotlib
